Moved generalEqual to expression.utils and change exception type
