+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/relation/base/local/BasicRegRelation.java
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.dat3m.dartagnan.program.utils.Utils.generalEqual;
+import static com.dat3m.dartagnan.expression.utils.Utils.generalEqual;
 
 abstract class BasicRegRelation extends StaticRelation {
