+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/relation/base/memory/RelLoc.java
@@ -2,7 +2,7 @@ package com.dat3m.dartagnan.wmm.relation.base.memory;
 
 import com.dat3m.dartagnan.program.event.Event;
 import com.dat3m.dartagnan.program.event.MemEvent;
-import com.dat3m.dartagnan.program.utils.Utils;
+import com.dat3m.dartagnan.expression.utils.Utils;
 import com.dat3m.dartagnan.wmm.filter.FilterBasic;
 import com.dat3m.dartagnan.wmm.relation.Relation;
 import com.dat3m.dartagnan.wmm.utils.Tuple;
